Relearning Datagen
Documented Code.
- Added some comments in the code to show where stuff should be added when new items/blocks are created.
- Added comments for where I will add stuff for Loot Pools in future commits

While looking in to how to make loot pools, I am trying to re-figure out how the Data Generation works in Minecraft.

From what I remember, and have found in testing:
- DataGeneration class is the entry point for all data generation methods. Will declare a "provider" of each type (Loot provider, Tag provider, Recipe provider, and Model provider)
- [type]Provider class defines what items/blocks/tags etc should have a .json file generated for it.

There are alot of "added" .json files in my commit, because I tested deleting them, and regenerating them though gradle's "runData" function instead of using "runClient" to run the code.
